+++ js/ui/react/App.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { StructureBrowser } from './components/StructureBrowser.js'
 import { SearchBar } from './components/SearchBar.js'
 import { InfoPanel } from './components/InfoPanel.js'
+import { MultiSelectPanel } from './components/MultiSelectPanel.js'
 
 export function App() {
   return (
@@ -9,6 +10,7 @@ export function App() {
       <SearchBar />
       <StructureBrowser />
       <InfoPanel />
+      <MultiSelectPanel />
     </>
   )
 }